Player Logic & Character Selection screens, assets
Files for player logic and character selection (dependent on CurrentCharacter.txt and MenuWidgets.py), as well as assets folder with updated sprites and sprite sheets. Player logic is currently runnable for demo (this will most likely be changed) and character selection screen currently closes on "OK" instead of bringing user to settings -- will add this functionality when all screens are being connected.